perf(stocks): lazy-load admin panel and modals on ticker page
The /stocks/[exchange]/[ticker] page eagerly bundled four sets of
client-side modals into the main JS chunk even though they're all
hidden until user interaction (or never visible to non-admins). This
extracted ~700+ lines of modal code plus the marked markdown parser
from the critical path, reducing main-thread script eval time.

Changes:
- StockActions: extract admin-only dropdown + 4 modals + parseMarkdown
  into new StockActionsAdminPanel, dynamically imported. Non-admin
  visitors never download this chunk.
- TickerComparisonButton: dynamic-import ComparisonModal (pulls in
  TickerComparison table + FullScreenModal); only mounts on first click.
- FavouriteButton: dynamic-import AddEditFavouriteModal,
  ManageListsModal, ManageTagsModal; only mount on first click.
- NotesButton: dynamic-import AddEditNotesModal; only mounts on first
  click.

Each lazy modal uses a `hasMountedModal` flag so the close animation
keeps working after the first open.
